+++ CHANGELOG.md
@@ -1,6 +1,7 @@
-## 1.x
+## 1.3.0
 
 - Tell the keyboard that the first letters of sentences should be capitalized in the prompt, verse text, and collection name TextFields.
+- Tap the words in the letter hints to make them visible.
 
 ## 1.2.1 - July 15, 2023